JOB_SEARCH_v1/requirements.txt
python-jobspy
pandas 
scikit-learn 
groq 
pdfplumber 
python-docx
python-dotenv
sentence-transformers
